refactor: TimetableRepository 벌크 DELETE 쿼리 적용 및 dayOfWeek 변환 방식 통일 (#33)

+++ src/main/java/com/cotato/blankit/domain/timetable/service/TimetableService.java
@@ -34,7 +34,7 @@ public class TimetableService {
     @Transactional
     public TimetableResponse createTimetable(Long userId, TimetableCreateRequest request) {
         validateTimeRange(request.startTime(), request.endTime());
-        checkTimeConflict(userId, (byte) request.dayOfWeek().intValue(),
+        checkTimeConflict(userId, request.dayOfWeek().byteValue(),
                 request.startTime(), request.endTime(), null);
 
         User user = getUser(userId);
